+++ database/factories/UserFactory.php
@@ -6,7 +6,9 @@ use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
 
-/** @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\User> */
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\User>
+ */
 class UserFactory extends Factory
 {
     /** The current password being used by the factory. */